PowerModelsSecurityConstrained Algorithmic Components

Data Wangling

These tools are used for parsing data files and translating data into the PowerModels internal data model.

Missing docstring.

Missing docstring for parse_goc_files. Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_pm_model. Check Documenter's build log for details.

OPF Formulations

These are standard OPF formulations (i.e. without contingency constraints) and are used as sub-processes in building contingency constrained solutions.

Missing docstring.

Missing docstring for run_opf_shunt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_opf_cheap. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_opf_cheap_dc. Check Documenter's build log for details.

PowerModelsSecurityConstrained.run_opf_pg_pf_rect_5 — Function

Contingency Filters

These are tools for checking for constraint violations in contingencies

PowerModelsSecurityConstrained.check_contingencies_branch_flow_ratec_nd_first_lazy — Function

given a network, checks the operating point against the contingencies to look for violations

SCOPF Formulations

These are contingency constrained OPF formulations.

PowerModelsSecurityConstrained.run_scopf_cuts_dc_soft_2 — Function

Simple DC SCOPF with flow cuts

Second-Stage Solvers

These are solvers for quickly solving large collections of contingencies. These solvers are usually more detailed on than the models used in the contingency filters.

Missing docstring.

Missing docstring for run_fixpoint_pf_v2_3!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for run_fixpoint_pf_soft!. Check Documenter's build log for details.