Solver Components

Data Wangling

These tools are used for parsing data files and translating data into the PowerModels internal data model.

PowerModelsSecurityConstrained.parse_c1_case — Function

Given an .ini file and a scenario id (i.e. a directory name), looks for and parses the files from ARPA-e GOC Challenge 1 data format.

.raw network model

.rop generator costs

.con contingency set

.inl generator contingency response parameters

PowerModelsSecurityConstrained.build_c1_pm_model — Function

transforms files from ARPA-e GOC Challenge 1 data format in to the PowerModels data format. This consists of taking the data from multiple data structures and putting into a network data dictionary.

PowerModelsSecurityConstrained.parse_c2_case — Function

Given an a directory path looks for and parses the files from ARPA-e GOC Challenge 2 data format.

.raw network model

.con contingency set

.json additional model data

PowerModelsSecurityConstrained.build_c2_pm_model — Function

Transforms files from ARPA-e GOC Challenge 2 data format in to the PowerModels data format. This consists of taking the data from multiple data structures and putting into a network data dictionary.

OPF Formulations

These are standard OPF formulations (i.e. without contingency constraints) and are used as sub-processes in building contingency constrained solutions.

PowerModelsSecurityConstrained.run_c1_opf_shunt — Function

An OPF formulation conforming to the ARPA-e GOC Challenge 1 specification. Power balance and branch flow constraints are strictly enforced. The primary departure from the PowerModels standard formulation is dispatchable bus shunts and a slight change in the transformer model.

PowerModelsSecurityConstrained.run_c1_opf_cheap — Function

An OPF formulation conforming to the ARPA-e GOC Challenge 1 specification. Power balance are strictly enforced and the branch flow violations are penalized based on a conservative linear approximation of the formulation's flow violation penalty specification.

PowerModelsSecurityConstrained.run_c1_opf_cheap_lazy_acr — Function

A variant of run_opf_cheap model, specialized for solving very large AC Power Flow models in rectangular coordinates for faster derivative computations. Support sparse collections of flow constrains for increased performance.

PowerModelsSecurityConstrained.build_c2_opf_soft — Function

PowerModelsSecurityConstrained.build_c2_opf_soft_ctg — Function

OPF with UC Formulations

These are standard OPF formulations (i.e. without contingency constraints) with discrete variables for supporting the commitment of generation units.

PowerModelsSecurityConstrained.build_c2_opf_uc — Function

OTS Formulations

These are standard transmission switching formulations (i.e. without contingency constraints) and are used to find economic improving topology changes in a network.

PowerModelsSecurityConstrained.build_c2_ots_soft — Function

Contingency Filters

These are tools for checking for constraint violations in contingencies

PowerModelsSecurityConstrained.check_c1_contingency_violations — Function

Checks a given operating point against the contingencies to look for branch flow violations. The DC Power Flow approximation is used for flow simulation. Returns a list of contingencies where a violation is found.

PowerModelsSecurityConstrained.check_c1_contingencies_branch_power — Function

Checks a given operating point against the contingencies to look for branch flow violations. The DC Power Flow approximation is used for flow simulation. If a violation is found, computes a PTDF cut based on bus injections. Uses the participation factor based generator response model from the ARPA-e GOC Challenge 1 specification.

PowerModelsSecurityConstrained.check_c1_contingencies_branch_power_bpv — Function

A variant of check_contingencies_branch_flow_remote, which ignores the participation factor based generator response model from the ARPA-e GOC Challenge 1 specification and instead injects active power equally at all buses in the network.

SCOPF Formulations

These are contingency constrained OPF formulations.

PowerModelsSecurityConstrained.run_c1_scopf — Function

An academic SCOPF formulation inspired by the ARPA-e GOC Challenge 1 specification. Power balance and line flow constraints are strictly enforced in the first stage and contingency stages.

This formulation is best used in conjunction with the contingency filters that find violated contingencies.

PowerModelsSecurityConstrained.run_c1_scopf_cuts_soft — Function

An SCOPF formulation conforming to the ARPA-e GOC Challenge 1 specification. A DC power flow approximation is used. Power balance and line flow constraints are strictly enforced in the first stage. Contingency branch flow constraints are enforced by PTDF cuts and penalized based on a conservative linear approximation of the formulation's specification.

This formulation is used in conjunction with the contingency filters that generate PTDF cuts.

PowerModelsSecurityConstrained.run_c1_scopf_cuts_soft_bpv — Function

a variant of run_c1_scopf_cuts_dc_soft with a different generator response function

PowerModelsSecurityConstrained.run_c1_scopf_contigency_cuts — Function

Solves a SCOPF problem by iteratively checking for violated contingencies and resolving until a fixed-point is reached

PowerModelsSecurityConstrained.run_c1_scopf_ptdf_cuts! — Function

Solves a SCOPF problem by iteratively checking for violated branch flow constraints in contingencies and resolving until a fixed-point is reached.

The base-case model is formulation agnostic. The flow cuts are based on PTDF and utilize the DC Power Flow assumption.

Contingency-Stage Solvers

These are solvers for quickly solving large collections of contingencies. These solvers are usually more detailed on than the models used in the contingency filters.

PowerModelsSecurityConstrained.run_c1_fixpoint_pf_pvpq! — Function

A power flow solver conforming to the ARPA-e GOC Challenge 1 contingency-stage specification. The solver conducts multiple rounds of PV/PQ bus switching to heuristically enforce the disjunctive constraints.

PowerModelsSecurityConstrained.run_c1_fixpoint_pf_bqv! — Function

A power flow solver inspired by the ARPA-e GOC Challenge 1 contingency-stage specification but designed to be faster on large network cases. Instead conducting PV/PQ bus switching address disjunctive constraints, this solver simply adds extra reactive capability on buses where the voltage bounds cannot be enforced and takes an associated power balance penalty.